fix: castear getCode() a int en AuthController para evitar TypeError 500
$e->getCode() puede retornar string en algunas excepciones de PHP.
errorResponse() exige int en el segundo argumento, lo que causaba un
TypeError que se manifestaba como HTTP 500 al hacer login desde el frontend.

+++ app/Modules/Auth/Controllers/AuthController.php
@@ -58,7 +58,7 @@ class AuthController extends Controller
             ], 'Inicio de sesión exitoso.');
 
         } catch (\Exception $e) {
-            return $this->errorResponse($e->getMessage(), $e->getCode() ?: 401);
+            return $this->errorResponse($e->getMessage(), (int) ($e->getCode() ?: 401));
         }
     }
